使用模板模式建立士兵、敌人的攻击与受击的框架

+++ Assets/Scripts/GameSystems/CharacterSystem/AttrStrategy/EnemyAttrStrategy.cs
@@ -8,7 +8,7 @@ namespace GameSystems.CharacterSystem.AttrStrategy
 
         public int GetDamageReduceValue(int level) => 0;
 
-        public int GetCriticalDamage(int criticalRate)
+        public int GetCriticalDamage(float criticalRate)
         {
             return Random.Range(0, 1f) <= criticalRate 
                 ? (int) (Random.Range(0.5f, 1f) * 10) : 0;
